Added input validation to company registration form
