+++ src/components/Job/ApplyBtn.js
@@ -1,55 +1,15 @@
-import React, { useState } from 'react';
+import React from 'react';
 import styled from 'styled-components/macro';
-import TextInput from '../TextInput';
 import Button from '../Button';
-
-import { db } from '../../firebase';
 
 const Container = styled.div`
   .apply-btn {
     margin: 0 0 10px 0;
   }
-
-  .email-collect {
-    display: block;
-    padding: 20px;
-    background-color: ${({ theme }) => theme.uiBackground};
-    border-radius: 6px;
-    text-align: center;
-
-    h4 {
-      margin: 0 0 10px 0;
-    }
-
-    .email-input {
-      margin-bottom: 10px;
-    }
-
-    .no-thanks {
-      margin-top: 10px;
-      font-size: 12px;
-      color: ${({ theme }) => theme.textMedium};
-      cursor: pointer;
-
-      &:hover {
-        color: ${({ theme }) => theme.textDark};
-        text-decoration: underline;
-      }
-    }
-  }
 `;
 
-const ApplyBtn = ({ job, job: { applyUrl, id } = {}, companyName = '' }) => {
-  const [showForm, setShowForm] = useState(false);
-  const [applicantEmail, setApplicantEmail] = useState('');
-  const [submitState, setSubmitState] = useState({
-    submitting: false,
-    error: false,
-    success: false
-  });
-
+const ApplyBtn = ({ job, job: { applyUrl } = {}, companyName = '' }) => {
   const onApplyClick = () => {
-    // setShowForm(true);
     window.ga('send', {
       hitType: 'event',
       eventCategory: 'Jobs',
@@ -59,50 +19,8 @@ const ApplyBtn = ({ job, job: { applyUrl, id } = {}, companyName = '' }) => {
     window.open(applyUrl);
   };
 
-  const onFormSubmit = e => {
-    e.preventDefault();
-    setSubmitState({ submitting: true, error: false, success: false });
-    db.collection('applications')
-      .add({
-        jobID: id,
-        applyUrl,
-        applicantEmail
-      })
-      .then(function() {
-        setSubmitState({ submitting: false, error: false, success: true });
-      })
-      .catch(function(error) {
-        setSubmitState({ submitting: false, error: error, success: false });
-      });
-    setTimeout(() => {
-      window.open(applyUrl);
-    }, 2000);
-    return false;
-  };
-
-  let content;
-  if (submitState.success) {
-    content = <p>Success! Taking you to the job application...</p>;
-  } else if (showForm) {
-    const submitLabel = submitState.submitting ? 'Submitting...' : 'Submit';
-    content = (
-      <form className="email-collect" onSubmit={onFormSubmit}>
-        <h4>Leave your email so we can stay in touch!</h4>
-        <TextInput
-          className="email-input"
-          placeholder="Email"
-          onChange={e => setApplicantEmail(e.target.value)}
-          value={applicantEmail}
-          fullWidth
-        />
-        <Button label={submitLabel} size="large" fullWidth type="submit" />
-        <a href={applyUrl} className="no-thanks" target="_blank">
-          No thanks, just take me to the application
-        </a>
-      </form>
-    );
-  } else {
-    content = (
+  return (
+    <Container>
       <Button
         className="apply-btn"
         label="Apply"
@@ -110,10 +28,8 @@ const ApplyBtn = ({ job, job: { applyUrl, id } = {}, companyName = '' }) => {
         onClick={onApplyClick}
         fullWidth
       />
-    );
-  }
-
-  return <Container>{content}</Container>;
+    </Container>
+  );
 };
 
 export default ApplyBtn;
